Change approach to gcloud function
